refactor(core): extract retrieve_candidates + BM25 helpers to rag_retrieval (#461)
Closes #461. Second slice (PR-H1b) of the rag_core.py post-retrieval
decomposition (PR-H1a was #459/#460).

Moves the candidate-generation + scoring-primitive surface (rag_core.py
lines 1753-2140) to rag_retrieval.py:

- retrieve_candidates — pre-fusion phase (filter + per-chunk dense /
  lexical / metadata / BM25 / m3 scoring). Mutates plan with
  candidate_count, total_chunks, filter_fallback_used.
- embed_query_for_index — backend-dispatched query embedding
  (sentence-transformers / openai / hashing fallback).
- dense_similarity, lexical_similarity, metadata_similarity — the
  three per-chunk scoring primitives.
- _strip_bm25_extra_suffixes, _apply_bm25_extra_filter,
  _chunk_tokens_for_bm25, get_or_build_bm25, bm25_scores_for_index —
  the BM25 corpus + scoring surface (issue #150 bm25_extra profile).

rag_retrieval.py grows from 267 to ~700 LOC, owning the full
retrieval pipeline. rag_core.py drops ~390 LOC (4,201 → 3,811).

Circular-import handling
------------------------
Module-level imports added to rag_retrieval (all safe — no path back
to rag_core):
- korean_lexicon (BM25_EXTRA_PARTICLE_SUFFIXES, BM25_EXTRA_STOPWORDS)
- rag_pipeline_presets (VALID_BM25_STOPWORD_PROFILES)
- rag_query_expansion (default_expander)
- rank_bm25 (BM25Okapi as _BM25Okapi, guarded)

Late-imports inside specific functions (rag_core symbols that serve
many non-retrieval call sites and stay there as canonical homes):
- retrieve_candidates: normalize_regions, normalize_page_span
- embed_query_for_index: DEFAULT_EMBEDDING_MODEL, DEFAULT_HASH_DIM,
  embed_texts, hashing_embeddings
- _chunk_tokens_for_bm25: tokenize
- lexical_similarity: tokenize

rag_core re-exports all 10 public names + _apply_bm25_extra_filter
(friend-of-module test export at tests/test_hybrid_retrieval_regression.py)
so every existing caller keeps working.

Cross-ref updates
-----------------
- ADR 0023 (HyDE): pointer changed from rag_core.retrieve_candidates
  to rag_retrieval.retrieve_candidates (3 locations) + line number
  L1780 annotated as the pre-extraction location.
- CLAUDE.md: rag_retrieval entry + rag_query_expansion entry point at
  the new module path.

Verification
------------
- Full local pytest 838 passed, 1 skipped, 0 failed.
- JSON-identity preserved: tests/test_naive_baseline_ranking_invariance.py
  (ADR 0001 golden) + tests/test_retrieval_loop_regression.py +
  tests/test_langgraph_orchestrator_regression.py all green.
